requirements.txt
asyncio==3.4.3
selenium==4.0.0
beautifulsoup4==4.10.0
aiogram>=2.14.2,<3.0.0
python-dotenv==1.0.0